backend/requirements.txt
# ========== Core Dependencies ==========
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# ========== Audio Processing ==========
# Audio Libraries
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
audioread==3.0.1
noisereduce==3.0.0

# Praat Python
praat-parselmouth==0.4.3

# ========== Speech Recognition ==========
# Whisper
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# Google Cloud STT (optional)
# google-cloud-speech==2.21.0

# Azure STT (optional)
# azure-cognitiveservices-speech==1.34.0

# ========== Korean Language Processing ==========
# Korean NLP
konlpy==0.6.0
jamo==0.4.1
JPype1==1.4.1  # Required for KoNLPy

# ========== Text Processing ==========
# String Matching
python-Levenshtein==0.23.0
# difflib2==0.1.0  # REMOVED - difflib is built-in Python library
rapidfuzz==3.5.2

# ========== Scientific Computing ==========
# Numerical Computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.3.2

# Signal Processing
matplotlib==3.8.2
seaborn==0.13.0

# ========== TextGrid ==========
# TextGrid Processing
textgrid==1.5
tgt==1.4.4

# ========== Utilities ==========
# File Handling
chardet==5.2.0
python-magic==0.4.27

# Async Support
aiofiles==23.2.1
# asyncio==3.4.3  # REMOVED - asyncio is built-in Python library

# Caching
diskcache==5.6.3

# Progress Bars
tqdm==4.66.1

# ========== Development Tools ==========
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Debugging
ipython==8.18.1
ipdb==0.13.13

# ========== Monitoring & Logging ==========
# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Monitoring (optional)
# prometheus-client==0.19.0
# sentry-sdk==1.38.0

# ========== API Documentation ==========
# OpenAPI
pydantic==2.5.2
pydantic-settings==2.1.0

# ========== Security ==========
# Authentication (optional)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# ========== Performance ==========
# Profiling
memory-profiler==0.61.0
line-profiler==4.1.1

# Parallel Processing
joblib==1.3.2
multiprocess==0.70.15

# ========== Machine Learning (Optional) ==========
# Deep Learning
# tensorflow==2.15.0
# transformers==4.35.2

# ========== Additional Korean STT (Optional) ==========
# Kakao STT
# kakao-sdk==1.0.0

# Naver Clova
requests==2.31.0

# ========== System Dependencies ==========
# Note: These may need to be installed separately via apt-get or brew
# - ffmpeg (for audio conversion)
# - libsndfile1 (for soundfile)
# - portaudio19-dev (for pyaudio)

# ========== Version Constraints ==========
# Python >= 3.8, < 3.12